# openapi-appengine.yaml
...
swagger: "2.0"
info:
  description: "IDC API Endpoints, version 1"
  title: "IDC API"
  version: "4.0"
#host: "api-dot-idc-dev.appspot.com"
host: "127.0.0.1:8095"
basePath: "/v1"
...
consumes:
- "application/json"
produces:
- "application/json"
schemes:
#- "https"
- "http"
...
paths:
  /about:
    get:
      description: "API Description and Swagger UI link"
      operationId: "root-get"
      produces:
        - "application/json"
      responses:
        200:
          description: "API Description and link to SwaggerUI interface."
          schema:
            $ref: "#/definitions/responseMessage"
        500:
          description: "Server error message"
          schema:
            $ref: "#/definitions/errorMessage"
  /swagger:
    get:
      description: "Swagger UI of the API spec"
      operationId: "swaggerUI"
      produces:
        - "text/html"
      responses:
        200:
          description: "Swagger UI of the API spec"
        500:
          description: "Server error message"
          schema:
            $ref: "#/definitions/errorMessage"
  /oauth2callback:
    get:
      description: "OAuth2 Redirect URI"
      operationId: "oauth2redirect"
      produces:
        - "text/html"
      responses:
        200:
          description: "OAuth2 Redirect URI"
        500:
          description: "Server error message"
          schema:
            $ref: "#/definitions/errorMessage"
  /collections:
    get:
      description: "Get a list of the available collections and versions of each."
      tags:
        - collections
      operationId: "getCollections"
      produces:
        - "application/json"
      responses:
        200:
          description: "A list of the collections available in this API with versions."
          schema:
            $ref: "#/definitions/collectionListResponse"
        500:
          description: "Server error message"
          schema:
            $ref: "#/definitions/errorMessage"

  # Different versions of a collection can have different fields
  /collections/{collection_id}/{version}:
    get:
      description: "Get list of the fields in a version of a collection"
      tags:
        - collections
      operationId: "getFields"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: collection_id
          in: path
          description: "Collection ID"
          type: string
          required: true
        - name: version
          in: path
          description: "Version"
          type: string
          required: true
      responses:
        200:
          description: "A list of the fields in a specific version of a collection"
          schema:
            $ref: "#/definitions/collectionFieldsResponse"
        500:
          description: "Server error message"
          schema:
            $ref: "#/definitions/errorMessage"
  /collections/{collection_id}/{version}/{field_name}:
    get:
      description: "Get metadata for a field in a version of a collection"
      tags:
        - collections
      operationId: "getFields"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: collection_id
          in: path
          description: "Collection ID"
          type: string
          required: true
        - name: version
          in: path
          description: "Version"
          type: string
          required: true
        - name: field_name
          in: path
          description: "Field Name"
          type: string
          required: true
      responses:
        200:
          description: "Metadata for a field in a collection"
          schema:
            $ref: "#/definitions/collectionFieldMetadataResponse"
        500:
          description: "Server error message"
          schema:
            $ref: "#/definitions/errorMessage"
...
definitions:
  responseMessage:
    type: "object"
    properties:
      message:
        type: "string"
      documentation:
        type: "string"
#      code:
#        type: "integer"
  collectionListResponse:
    type: "array"
    items:
      type: "object"
      properties:
        name:
          type: "string"
          $ref: "#/definitions/collectionNames"
        description:
          type: "string"
        versions:
          type: "array"
          items:
            type: "object"
            properties:
              version:
                type: "string"
              release_date:
                type: "date"
  collectionFieldsResponse:
    type: "array"
    items:
      type: "string"
  collectionFieldMetadataResponse:
    type: "object"
    properties:
      fieldName:
        type: "string"
      description:
        type: "string"
      fieldType:
        type: "string"
        enum: ["string", "date", "time", "timestamp", "integer", "float", "record", "array"]
      enumerated_values:
        type: "array"
        items:
          type: "string"
...
  collectionNames:
    type: "string"
    enum: [
      "tcga_blca",
      "tcga_brca",
      "tcga_cesc",
      "tcga_coad",
      "tcga_esca",
      "tcga_gbm",
      "tcga_hnsc",
      "tcga_kich",
      "tcga_kirc",
      "tcga_kirp",
      "tcga_lgg",
      "tcga_luad",
      "tcga_kusc",
      "tcga_ov",
      "tcga_prad",
      "tcga_reas",
      "tcga_sarc",
      "tcga_stad",
      "tcga_thca",
      "tcga_ucec"
      ]
...
  errorMessage:
    type: "object"
    properties:
      message:
        type: "string"
      code:
        type: "integer"
      not_found:
        type: "array"
        items:
          type: "string"
...
securityDefinitions:
#  # This section configures basic authentication with an API key.
#  api_key:
#    type: "apiKey"
#    name: "key"
#    in: "query"
## [END securityDef]
#  # This section configures authentication using Google API Service Accounts
#  # to sign a json web token. This is mostly used for server-to-server
#  # communication.
#  google_jwt:
#    authorizationUrl: ""
#    flow: "implicit"
#    type: "oauth2"
#    # This must match the 'iss' field in the JWT.
#    x-google-issuer: "jwt-client.endpoints.sample.google.com"
#    # Update this with your service account's email address.
#    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/jwk/YOUR-SERVICE-ACCOUNT-EMAIL"
#    # This must match the "aud" field in the JWT. You can add multiple audiences to accept JWTs from multiple clients.
#    x-google-audiences: "echo.endpoints.sample.google.com"
#  # This section configures authentication using Google App Engine default
#  # service account to sign a json web token. This is mostly used for
#  # server-to-server communication.
#  gae_default_service_account:
#    authorizationUrl: ""
#    flow: "implicit"
#    type: "oauth2"
#    # Replace YOUR-CLIENT-PROJECT-ID with your client project ID.
#    x-google-issuer: "YOUR-CLIENT-PROJECT-ID@appspot.gserviceaccount.com"
#    # Replace YOUR-CLIENT-PROJECT-ID with your client project ID.
#    x-google-jwks_uri: "https://www.googleapis.com/robot/v1/metadata/x509/YOUR-CLIENT-PROJECT-ID@appspot.gserviceaccount.com"
#    # This must match the "aud" field in the JWT. You can add multiple audiences to accept JWTs from multiple clients.
#    x-google-audiences: "echo.endpoints.sample.google.com"
#  # This section configures authentication using a service account
#  # to sign a json web token. This is mostly used for server-to-server
#  # communication.
#  google_service_account:
#    authorizationUrl: ""
#    flow: "implicit"
#    type: "oauth2"
#    # Replace YOUR-SERVICE-ACCOUNT-EMAIL with your service account email.
...